% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictWFH.R
\name{PredictWFH}
\alias{PredictWFH}
\title{Main module function to predict Occupation and Work from Home for each worker}
\usage{
PredictWFH(L)
}
\arguments{
\item{L}{A list containing the components listed in the Get specifications
for the module.}
}
\value{
A list containing the components specified in the Set
specifications for the module.
}
\description{
\code{PredictWFH} predicts Occupation and Work from Home for each worker.
It uses the model objects in data/OccShares_df.rda and data/WorkFromHome_df.rda.
}
