% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VETravelDemandWFH.R
\docType{_PACKAGE}
\name{VETravelDemandWFH}
\alias{VETravelDemandWFH}
\title{\code{VETravelDemandWFH} package}
\description{
Simulate Multi-Modal Travel Demand for Households with Adjustments for Working from Home
}
\details{
See the README on
\href{https://github.com/gregorbj/VisionEval/sources/modules/VETravelDemandWFH}{GitHub}
}
